Refactor HeaderMenu config file
